Rename assembler to assemblers
