+++ indra/assemblers/sbgn_assembler.py
