+++ indra/benchmarks/test_trips.py
@@ -1,5 +1,5 @@
-from indra.pysb_assembler import PysbAssembler
-from indra.trips import trips_api
+from indra.assemblers import PysbAssembler
+from indra import trips
 from os.path import dirname, join
 import indra.statements
 import sys
@@ -16,7 +16,7 @@ BRAF that is bound to NRAS and Vemurafenib binds NRAS-bound BRAF that is not bou
 def test_bind():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'The receptor tyrosine kinase EGFR binds the growth factor ligand EGF.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -28,7 +28,7 @@ def test_bind():
 def test_complex_bind():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'The EGFR-EGF complex binds another EGFR-EGF complex.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -42,7 +42,7 @@ def test_complex_bind():
 def test_complex_bind2():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'The EGFR-EGFR complex binds GRB2.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -56,7 +56,7 @@ def test_complex_bind2():
 def test_complex_bind3():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'RAF binds to the RAS-GTP complex.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -70,7 +70,7 @@ def test_complex_bind3():
 def test_complex_bind4():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'The RAF-RAS complex binds another RAF-RAS complex.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -86,7 +86,7 @@ def test_complex_bind4():
 def test_bound_mod():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'The adaptor protein GRB2 can bind EGFR that is phosphorylated on tyrosine.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -101,7 +101,7 @@ def test_bound_mod():
 def test_not_bound_to():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'BRAF that is not bound to Vemurafenib binds MEK1.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -118,7 +118,7 @@ def test_not_bound_to():
 def test_not_bound_to2():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'BRAF, not bound to Vemurafenib, phosphorylates MEK1.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_phosphorylation(st))
@@ -136,7 +136,7 @@ def test_not_bound_to2():
 def test_not_bound_to3():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'SOS1 bound to GRB2 binds NRAS that is not bound to BRAF and GTP.' 
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -150,7 +150,7 @@ def test_not_bound_to4():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'BRAF that is not bound to NRAS and Vemurafenib binds BRAF ' +\
           'that is not bound to Vemurafenib.' 
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -163,7 +163,7 @@ def test_not_bound_to4():
 def test_bound_to():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'NRAS, bound to GTP, binds BRAF.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -179,7 +179,7 @@ def test_bound_to():
 def test_bound_to2():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'EGFR-bound GRB2 binds SOS1.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -195,7 +195,7 @@ def test_bound_to2():
 def test_bound_to3():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'SOS1, bound to GRB2 binds RAS.' 
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -210,7 +210,7 @@ def test_bound_to3():
 def test_bound_to4(): 
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'RAS, bound to SOS1, binds GTP.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -225,7 +225,7 @@ def test_bound_to5():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'BRAF that is bound to NRAS and Vemurafenib binds' +\
           'BRAF that is bound to NRAS and Vemurafenib.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_complex(st))
@@ -239,7 +239,7 @@ def test_bound_to5():
 def test_transphosphorylate():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'EGFR, bound to EGFR, transphosphorylates itself.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_transphosphorylation(st))
@@ -254,7 +254,7 @@ def test_transphosphorylate():
 def test_transphosphorylate2():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'EGFR, bound to EGFR, transphosphorylates itself on a tyrosine residue.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_transphosphorylation(st))
@@ -269,7 +269,7 @@ def test_transphosphorylate2():
 def test_act_mod():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'MEK1, phosphorylated at Ser218 and Ser222, is activated.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_actmod(st))
@@ -285,7 +285,7 @@ def test_act_mod():
 def test_bound_phosphorylate():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'RAF, bound to RAF, phosphorylates MEK1.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_phosphorylation(st))
@@ -300,7 +300,7 @@ def test_bound_phosphorylate():
 def test_bound_not_bound_phosphorylate():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'BRAF-bound BRAF that is not bound to Vemurafenib phosphorylates MEK1.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_phosphorylation(st))
@@ -315,7 +315,7 @@ def test_bound_not_bound_phosphorylate():
 def test_act_phosphorylate():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'Active MEK1 phosphorylates ERK2 at Tyr187.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_phosphorylation(st))
@@ -330,7 +330,7 @@ def test_act_phosphorylate():
 def test_dephosphorylate():
     fname = sys._getframe().f_code.co_name + '.xml'
     txt = 'DUSP6 dephosphorylates ERK2 at Tyr187.'
-    tp = trips_api.process_text(txt, fname, False)
+    tp = trips.process_text(txt, fname, False)
     assert(len(tp.statements) == 1)
     st = tp.statements[0]
     assert(is_dephosphorylation(st))
