+++ indra/tests/test_biopax.py
@@ -2,7 +2,7 @@ import os
 from indra.java_vm import autoclass, cast
 from indra import biopax
 import indra.biopax.processor as bpc
-from indra.assembler import PysbAssembler
+from indra.assemblers import PysbAssembler
 
 model_path = os.path.dirname(os.path.abspath(__file__)) +\
              '/../../data/biopax_test.owl'
