+++ indra/tests/test_english_assembler.py
@@ -1,4 +1,4 @@
-import indra.assembler.english_assembler as ea
+import indra.assemblers.english_assembler as ea
 from indra.statements import *
 
 def test_agent_basic():
